:lipstick: Update request item component
