+++ src/components/Request/molecules/RequestList/index.tsx
@@ -1,5 +1,5 @@
 import * as S from './style'
-import MGMTListItem from 'components/common/atoms/MGMTListItem'
+import MGMTListItem from 'components/Request/atoms/RequestItem'
 export default function RequestList() {
   return (
     <S.RequestListWrapper>
@@ -10,6 +10,7 @@ export default function RequestList() {
         tagVisible={true}
         student="손정민"
         period="2023.05.26 ~ 2023.06.26"
+        id={1}
       />
       <MGMTListItem
         thumbnail="s"
@@ -18,6 +19,7 @@ export default function RequestList() {
         tagVisible={true}
         student="손정민"
         period="2023.05.26 ~ 2023.06.26"
+        id={2}
       />
       <MGMTListItem
         thumbnail="s"
@@ -26,6 +28,7 @@ export default function RequestList() {
         tagVisible={true}
         student="손정민"
         period="2023.05.26 ~ 2023.06.26"
+        id={2}
       />
     </S.RequestListWrapper>
   )
